Code refactoring technique apaplied
