+++ spring-data-relational/src/main/java/org/springframework/data/relational/core/dialect/AnsiDialect.java
@@ -84,7 +84,6 @@ public class AnsiDialect extends AbstractDialect {
 		return LOCK_CLAUSE;
 	}
 
-	@Override
 	public ArrayColumns getArraySupport() {
 		return ARRAY_COLUMNS;
 	}
